requirements.txt
canonicalwebteam.flask-base==2.5.0
